// src/types/generated/rocfinityV6.ts
export interface DefaultMintParams_CreateToken {
  __kind: 'CreateToken'
  tokenId: bigint
  initialSupply: bigint
  unitPrice: bigint
  cap: (TokenCap | undefined)
  behavior: (TokenMarketBehavior | undefined)
  listingForbidden: boolean
}
// ...
export type TokenMarketBehavior = TokenMarketBehavior_HasRoyalty | TokenMarketBehavior_IsCurrency

export interface TokenMarketBehavior_HasRoyalty {
  __kind: 'HasRoyalty'
  value: DefaultRoyalty
}

export interface TokenMarketBehavior_IsCurrency {
  __kind: 'IsCurrency'
}

export type TokenCap = TokenCap_SingleMint | TokenCap_Supply

export interface TokenCap_SingleMint {
  __kind: 'SingleMint'
}

export interface TokenCap_Supply {
  __kind: 'Supply'
  value: bigint
}